backend/requirements.txt
alembic==1.8.1
annotated-types==0.6.0
anyio==4.2.0
asyncpg==0.29.0
bcrypt==3.2.0
bzip2==1.0.8
ca-certificates==2024.3.11
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
expat==2.6.2
fastapi==0.103.0
greenlet==3.0.1
h11==0.14.0
idna==3.4
libffi==3.4.4
mako==1.2.3
markupsafe==2.1.3
openssl==3.0.13
passlib==1.7.4
pip==23.3.1
pydantic==2.5.3
pydantic-core==2.14.6
python==3.12.3
python-dateutil==2.8.2
python-dotenv==1.0.1
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
sqlalchemy==2.0.25
sqlite==3.41.2
starlette==0.27.0
tk==8.6.12
typing-extensions==4.9.0
typing_extensions==4.9.0
tzdata==2024a
uvicorn==0.20.0
vc==14.2
vs2015_runtime==14.27.29016
wheel==0.41.2
xz==5.4.6
zlib==1.2.13